Changed all data types relative to the amount of a single item to word (before we had int, uint, word...). Made CItem::GetAmount() and CItem::SetAmount() virtual, to avoid possible bugs when using CItemSpawn (which overrides those methods). Renamed CGRect class to CRect. Minor changes to CValueCurveDef::GetLinear.

+++ src/common/CExpression.cpp
@@ -294,7 +294,7 @@ int32 Calc_GetRandVal( int32 iQty )
 	if ( iQty <= 0 )
 		return 0;
 	if ( iQty >= INT32_MAX )
-		return ( MulDivLL(CSRand::genRandInt32(0, iQty), (uint)iQty, INT32_MAX ) );
+		return ( MulDivLL(CSRand::genRandInt32(0, iQty), iQty, INT32_MAX ) );
 	return CSRand::genRandInt32(0, iQty);
 }
